test/requirements.txt
pytest==8.4.2
cocotb==2.0.0
pycryptodome